updating shop items
