+++ routes/api/products.js
@@ -2,8 +2,8 @@ const productsController = require('../../controllers/productsController');
 
 module.exports = (app) =>{
   // FIXME: /api/createProduct or /api/products
-  app.post('/api/products',productsController.createProduct);
-  app.get("/api/findAll", productsController.findAll);
+  app.post('/api/products',productsController.create);
+  app.get("/api/products", productsController.getAllProducts);
   
   app.get("api/findbyid", productsController.findById);
