+++ server.js
@@ -29,6 +29,7 @@ app.use(passport.session()) // calls serializeUser and deserializeUser
 require("./routes/api/user")(app);
 require("./routes/api/job")(app);
 require("./routes/api/children")(app);
+require("./routes/api/products")(app);
 // FIXME: store route
